refactor: `domain` 패키지 분리 `service` service 패키지 추가
