+++ src/main/java/step3/service/Randomize.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3.service;
 
 import java.util.Random;
